Split up gulpfile
