+++ gulpfile.js/tasks/dev.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+var gulpSequence = require('gulp-sequence');
+
+function devTask() {
+	gulpSequence('server', 'default');
+}
+
+gulp.task('dev', devTask);
+module.exports = devTask;
